rename: Following Memory Log 디렉토리 네이밍 컨벤션에 맞게 수정 (#265)
