--- module-domain/src/main/java/com/depromeet/followingLog/port/in/command/CreateFollowingMemoryCommand.java
@@ -1,5 +1,0 @@
-package com.depromeet.followingLog.port.in.command;
-
-import com.depromeet.memory.domain.Memory;
-
-public record CreateFollowingMemoryCommand(Memory memory) {}
