+++ module-domain/src/main/java/com/depromeet/followinglog/port/in/FollowingMemoryLogUseCase.java
@@ -1,0 +1,10 @@
+package com.depromeet.followinglog.port.in;
+
+import com.depromeet.followinglog.domain.vo.FollowingLogSlice;
+import com.depromeet.followinglog.port.in.command.CreateFollowingMemoryCommand;
+
+public interface FollowingMemoryLogUseCase {
+    void save(CreateFollowingMemoryCommand followingMemoryCommand);
+
+    FollowingLogSlice findLogsByMemberIdAndCursorId(Long memberId, Long cursorId);
+}
