+++ module-domain/src/main/java/com/depromeet/followinglog/port/in/command/CreateFollowingMemoryCommand.java
@@ -1,0 +1,5 @@
+package com.depromeet.followinglog.port.in.command;
+
+import com.depromeet.memory.domain.Memory;
+
+public record CreateFollowingMemoryCommand(Memory memory) {}
