--- module-infrastructure/persistence-database/src/main/java/com/depromeet/followingLog/repository/FollowingMemoryLogJpaRepository.java
@@ -1,7 +1,0 @@
-package com.depromeet.followingLog.repository;
-
-import com.depromeet.followingLog.entity.FollowingMemoryLogEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FollowingMemoryLogJpaRepository
-        extends JpaRepository<FollowingMemoryLogEntity, Long> {}
