+++ module-infrastructure/persistence-database/src/main/java/com/depromeet/followinglog/repository/FollowingMemoryLogJpaRepository.java
@@ -1,0 +1,7 @@
+package com.depromeet.followinglog.repository;
+
+import com.depromeet.followinglog.entity.FollowingMemoryLogEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FollowingMemoryLogJpaRepository
+        extends JpaRepository<FollowingMemoryLogEntity, Long> {}
